requirements.txt
playwright>=1.43.0
selenium>=4.16.0
allure-pytest>=2.13.2
np~=1.0.2
allure-python-commons~=2.15.0